move dataproc expansion files to separate directory
